[ODOO-4491][ADD]sale_order_modifications] [UPD][sale_contact_partner]Rename and relocate fields and files to make possible the contacts zone to display in the sale orders [UPD][sale_order_workflow]Relocated and renamed the sale.order.zone file fo contact.zone in the module sale_contact_partner
