+++ sale_contact_partner/i18n/es_AR.po
@@ -6,8 +6,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 11.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2022-08-11 17:29+0000\n"
-"PO-Revision-Date: 2022-08-11 17:29+0000\n"
+"POT-Creation-Date: 2022-08-18 17:12+0000\n"
+"PO-Revision-Date: 2022-08-18 17:12+0000\n"
 "Last-Translator: <>\n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
@@ -21,18 +21,57 @@ msgid "Contact"
 msgstr "Contacto"
 
 #. module: sale_contact_partner
+#: model:ir.model.fields,field_description:sale_contact_partner.field_partner_zone_create_uid
+msgid "Created by"
+msgstr "Creado por"
+
+#. module: sale_contact_partner
+#: model:ir.model.fields,field_description:sale_contact_partner.field_partner_zone_create_date
+msgid "Created on"
+msgstr "Creado el"
+
+#. module: sale_contact_partner
+#: model:ir.model.fields,field_description:sale_contact_partner.field_partner_zone_display_name
+msgid "Display Name"
+msgstr "Nombre a Mostrar"
+
+#. module: sale_contact_partner
 #: code:addons/sale_contact_partner/models/res_partner.py:15
 #, python-format
 msgid "Email cannot have spaces ' '"
 msgstr "El correo no puede tener espacios' '"
 
 #. module: sale_contact_partner
-#: model:ir.model.fields,field_description:sale_contact_partner.field_sale_order_partner_zone
-msgid "Name"
-msgstr "Nombre"
+#: model:ir.model.fields,field_description:sale_contact_partner.field_partner_zone_id
+msgid "ID"
+msgstr "ID"
 
 #. module: sale_contact_partner
-#: model:ir.ui.view,arch_db:sale_contact_partner.sale_order_contact_form
+#: model:ir.model.fields,field_description:sale_contact_partner.field_partner_zone___last_update
+msgid "Last Modified on"
+msgstr "Última Modificación en"
+
+#. module: sale_contact_partner
+#: model:ir.model.fields,field_description:sale_contact_partner.field_partner_zone_write_uid
+msgid "Last Updated by"
+msgstr "Última Actualización por"
+
+#. module: sale_contact_partner
+#: model:ir.model.fields,field_description:sale_contact_partner.field_partner_zone_write_date
+msgid "Last Updated on"
+msgstr "Última Actualización el"
+
+#. module: sale_contact_partner
+#: model:ir.model.fields,field_description:sale_contact_partner.field_partner_zone_name
+#: model:ir.model.fields,field_description:sale_contact_partner.field_sale_order_partner_zone
+#: model:ir.model.fields,field_description:sale_contact_partner.field_sale_order_partner_zone_ids_name
+msgid "Name"
+msgstr "Zona"
+
+#. module: sale_contact_partner
+#: model:ir.model.fields,field_description:sale_contact_partner.field_res_partner_partner_zone_ids
+#: model:ir.model.fields,field_description:sale_contact_partner.field_res_users_partner_zone_ids
+#: model:ir.model.fields,field_description:sale_contact_partner.field_sale_order_partner_zone_ids
 msgid "Partner Zone"
 msgstr "Zona"
 
@@ -48,8 +87,18 @@ msgid "Quotation"
 msgstr "Presupuesto"
 
 #. module: sale_contact_partner
-#: model:ir.model.fields,field_description:sale_contact_partner.field_res_partner_sale_zone_ids
-#: model:ir.model.fields,field_description:sale_contact_partner.field_res_users_sale_zone_ids
-msgid "Sale Zone"
+#: model:ir.model,name:sale_contact_partner.model_sale_report
+msgid "Sales Orders Statistics"
+msgstr "Estadísticas pedidos de ventas"
+
+#. module: sale_contact_partner
+#: model:ir.model.fields,field_description:sale_contact_partner.field_sale_report_partner_zone_ids
+#: model:ir.ui.view,arch_db:sale_contact_partner.view_sale_order_zone_filter
+msgid "Zone"
 msgstr "Zona"
+
+#. module: sale_contact_partner
+#: model:ir.model,name:sale_contact_partner.model_partner_zone
+msgid "partner.zone"
+msgstr "partner.zone"
