+++ sale_contact_partner/models/__init__.py
@@ -1,2 +1,4 @@
 from . import res_partner
 from . import sale_order
+from . import partner_zone
+from . import sale_report
