+++ sale_contact_partner/models/partner_zone.py
@@ -1,0 +1,10 @@
+from odoo import models, fields, api
+
+
+class PartnerZone(models.Model):
+    _name = "partner.zone"
+
+
+    name = fields.Char(required=True)
+    description = fields.Char
+
