+++ sale_contact_partner/security/ir.model.access.csv
@@ -1,0 +1,2 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_partner_zone,access_partner_zone,model_partner_zone,base.group_user,1,1,1,1
