+++ sale_contact_partner/views/partner_zone_data.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <odoo noupdate="1">
-    <record model='sale.order.zone' id='zone1'>
+    <record model='partner.zone' id='zone1'>
         <field name='name'>Zona 1</field>
     </record>
-    <record model='sale.order.zone' id='zone2'>
+    <record model='partner.zone' id='zone2'>
         <field name='name'>Zona 2</field>
     </record>
-    <record model='sale.order.zone' id='zone3'>
+    <record model='partner.zone' id='zone3'>
         <field name='name'>Zona 3</field>
     </record>
 </odoo>
