+++ sale_contact_partner/views/sale_order_form_inherited_view.xml
@@ -1,13 +1,28 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="sale_order_contact_form" model="ir.ui.view">
-        <field name="name">sale.order.contact.inherited.form</field>
-        <field name="model">sale.order</field>
-        <field name="inherit_id" ref="sale_order_workflow.sale_order_form_design"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='picking_status']" position="after">
-                <field name="partner_zone" string="Partner Zone" readonly="1" force_save="1"/>
-            </xpath>
-        </field>
+
+    <record id="sale_order_form_design" model="ir.ui.view">
+      <field name="model">sale.order</field>
+      <field name="name">Sale Order Desgin</field>
+      <field name="inherit_id" ref="sale.view_order_form" />
+      <field eval="100" name="priority"/>
+      <field name="arch" type="xml">
+        <xpath expr="//field[@name='team_id']" position="after">
+            <field name="partner_zone" readonly="1"/>
+        </xpath>
+      </field>
     </record>
+
+    <record id="view_sale_order_zone_filter" model="ir.ui.view">
+      <field name="name">sale.order.zone.filter</field>
+      <field name="model">sale.order</field>
+      <field name="inherit_id" ref="sale.sale_order_view_search_inherit_quotation"/>
+      <field name="arch" type="xml">
+        <xpath expr="//group[1]" position="inside">
+          <filter name="groupby_partner_zone" string="Zone" context="{'group_by': 'partner_zone'}"/>
+        </xpath>
+      </field>
+    </record>
+
+
 </odoo>
