+++ sale_order_modifications/i18n/es_AR.po
@@ -6,8 +6,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 11.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2022-08-12 20:15+0000\n"
-"PO-Revision-Date: 2022-08-12 20:15+0000\n"
+"POT-Creation-Date: 2022-08-18 17:11+0000\n"
+"PO-Revision-Date: 2022-08-18 17:11+0000\n"
 "Last-Translator: <>\n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
@@ -17,6 +17,6 @@ msgstr ""
 
 #. module: sale_order_modifications
 #: model:ir.ui.view,arch_db:sale_order_modifications.sale_order_form_new_fields
-msgid "Quotation Zone"
-msgstr "Zona de Presupuesto"
+msgid "Partner Zone"
+msgstr "Zona"
