+++ sale_order_workflow/models/__init__.py
@@ -3,5 +3,3 @@
 from . import sale_order
 from . import mrp_production
 from . import stock_picking
-from . import sale_order_zone
-from . import sale_report
