+++ sale_order_workflow/models/sale_order.py
@@ -10,9 +10,6 @@ import dateutil.parser
 class SaleOrder(models.Model):
     _inherit = "sale.order"
 
-    zone_ids = fields.Many2one("sale.order.zone", string="Zone")
-    zone_ids_name = fields.Char(related="zone_ids.name", store=True)
-
     state = fields.Selection(
         [
             ("draft", "Quotation"),
