--- sale_order_workflow/models/sale_order_zone.py
@@ -1,9 +1,0 @@
-from odoo import models, fields, api
-
-
-class SaleZone(models.Model):
-    _name="sale.order.zone"
-
-    name= fields.Char()
-    description = fields.Char()
-
